pyproject.toml
[tool.poetry]
name = "foundry-world-tools"
version = "0.4.8"
description = "Foundry World Tools"
authors = ["Alastair Marchant <alastairmarchant.am@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/alastairmarchant/foundry-world-tools"
repository = "https://github.com/alastairmarchant/foundry-world-tools"
documentation = "https://foundry-world-tools.readthedocs.io"
packages = [
    { include = "fwt", from = "src" },
]
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/alastairmarchant/foundry-world-tools/releases"

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.0.1"
jsonlines = ">=3.1.0"
pyyaml = ">=6.0.1"
importlib-resources = ">=6.0.1"
typing-extensions = "^4.7.1"

[tool.poetry.group.pre-commit]
optional = true

[tool.poetry.group.pre-commit.dependencies]
black = ">=21.10b0"
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
isort = ">=5.10.1"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pyupgrade = ">=2.29.1"

[tool.poetry.group.safety]
optional = true

[tool.poetry.group.safety.dependencies]
safety = ">=1.10.3"

[tool.poetry.group.coverage]
optional = true

[tool.poetry.group.coverage.dependencies]
coverage = {extras = ["toml"], version = ">=6.2"}

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
Pygments = ">=2.10.0"
pytest = ">=6.2.5"

[tool.poetry.group.mypy]
optional = true

[tool.poetry.group.mypy.dependencies]
mypy = ">=0.930"
types-pyyaml = "^6.0.12.11"

[tool.poetry.group.typeguard]
optional = true

[tool.poetry.group.typeguard.dependencies]
typeguard = ">=2.13.3"

[tool.poetry.group.xdoctest]
optional = true

[tool.poetry.group.xdoctest.dependencies]
xdoctest = {extras = ["colors"], version = ">=0.15.10"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
furo = ">=2021.11.12"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
fwt = "fwt.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["fwt", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
hide_error_codes = false
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
